Yet more mucking about with the description for Apple Podcasts.
Also took the chance to fix links in the show notes that contain
apostrophes for Spotify.

Took the opportunity to update the image for the Robbit Mon Dieu episode so that it's large enough for Apple Podcasts's picky standards, as well.

I also lowered the quality of the logo image a bit â hopefully
this'll stop Apple Podcasts from going mad and compressing it to heck
and back? Probably not. Hope is the last thing to die, or however that
goes.