Added Spotify button.
I don't like that they've been making podcasts exclusive to their platform – podcasts are decentralized for a reason – but it has to be said that they've got a darn good podcast app. The only one that supports chapters that I've found! (Except for the podcast player I'm using on the episode pages!)