Removed Google Podcasts button; added Pocket Casts.
Since Google is putting the kibosh on Google Podcasts! Welp!

Pocket Casts is entirely a self-centered choice – and not based on any reasonable usage statistics – since it's the app I use. :^)

I also added support for a YouTube Music button, if at some point in the future Google should get that working (and I can stomach the situation).