Added video game title to episode titles.
I don't necessarily feel like this is needed, but it might help listeners pick an episode they're particularly interested, and – more importantly – at least some podcast apps (Spotify…) don't index the show notes when searching for episodes. Thus, the game's title must be in the episode title for it to be produced as a result when searching.

The way the game title is added is currently after a thick space, an em dash, and another thick space – I considered a wealth of different brackets so that a title containing an en dash wouldn't look too confusing, but they all looked tacky or busy to various degrees. I'll just live with the fact that having an en dash in an episode title might look a bit discombobulating.

Also in this commit, I took the liberty of replacing the copyright symbol in the copyright notice, because I realized that in this day and age, a copyright notice serves no purpose but to assert to the consumer that "I own this, and you'll be in trouble if you copy it!" That's not a message I want to send. That's tacky and obnoxious. What did I replace it with? The anachronistic "sound recording copyright symbol" (P for "phonograph"), which I have never seen in my life. It's apparently still codified into US law in some manner, however…? Seemingly? The irreverence and mystifying nature of this symbol negates any disagreeable aura a copyright notice might otherwise exude, I feel.