Autofill file size and duration into post's source files directly.
Octopod had an approach where it would recalculate them every time the site was built if they weren't filled out – but I don't exactly like the idea of that. Much better to fill it out for good at build time. Perhaps this could be integrated into VS code or somethin' so one doesn't need to rebuild the site locally every time a new episode is to be uploaded (although I certainly don't see myself *not* doing that anytime soon) – or perhaps even a Git commit hook?

This isn't the first time I find myself in the unenviable position of writing code to edit YAML with the desire to preserve whitespace and comments… When was the last time again? I seem to recall it being in Crystal… Which, I suppose, means I've already done more or less this exact thing in a… Ruby-*like?* What a guy, man. What a guy.

I should still implement reading chapters from the audio files – but
as it stands, this is not a feature in either of the two semi-large
Ruby audio metadata libraries. I foresee a funky future where I end up
implementin' it myself and pullin' a request… That'd be fun!?