Apparently, "and" and "or" differ from "&&" and "||" in Ruby!
And you're practically only supposed to use the latter pair? And the situation with "not" and "!" is similar? What a bizarre design relic. I dream of them changing this in Ruby 4. Though judging from experience with Ruby developers, they seem to not infrequently be fairly conservative about the quirks of language – even the less pleasant ones, hahah.

I can't believe I didn't find out about this until now, though!?